add generics exercises 2

+++ Generics_Exercises/armazem_especifico/Comida.java
@@ -1,16 +1,16 @@
 public class Comida {
     private String nome;
-    private int calorias;
+    private double calorias;
     private double preco;
 
-    public Comida(String nome, int calorias, double preco) {
+    public Comida(String nome, double calorias, double preco) {
         this.nome = nome;
         this.calorias = calorias;
         this.preco = preco;
     }
 
     public String getNome() { return nome; }
-    public int getCalorias() { return calorias; }
+    public double getCalorias() { return calorias; }
     public double getPreco() { return preco; }
 
     @Override
